# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoluStore API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: settings
    description: Store settings
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: List featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /categories:
    get:
      operationId: listCategories
      tags: [products]
      summary: List active categories with product counts
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"

  /store-stats:
    get:
      operationId: getStoreStats
      tags: [products]
      summary: Get store statistics
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /store-settings:
    get:
      operationId: getStoreSettings
      tags: [settings]
      summary: Get public store settings (name, logo, tagline)
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get all store settings (admin)
      responses:
        "200":
          description: Full store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
        "401":
          description: Not authenticated
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
        "401":
          description: Not authenticated

  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories (admin)
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          description: Not authenticated
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Not authenticated

  /admin/categories/{id}:
    patch:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Not authenticated
        "404":
          description: Not found
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Not found

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin list all products
      responses:
        "200":
          description: All products including inactive
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated

  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update product fields including price and stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Not authenticated
        "404":
          description: Not found
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, description, category, price, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        stock:
          type: integer
        badge:
          type: ["string", "null"]
          description: HOT, NEW, SALE, or LIMITED
        imageUrl:
          type: ["string", "null"]
        buyLink:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, description, category, price, stock]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        price:
          type: number
          minimum: 0
        originalPrice:
          type: number
        stock:
          type: integer
          minimum: 0
        badge:
          type: string
        imageUrl:
          type: string
        buyLink:
          type: string
        isActive:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        price:
          type: number
          minimum: 0
        originalPrice:
          type: ["number", "null"]
        stock:
          type: integer
          minimum: 0
        badge:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        buyLink:
          type: ["string", "null"]
        isActive:
          type: boolean

    Category:
      type: object
      required: [id, slug, label, icon, isActive, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        label:
          type: string
        icon:
          type: string
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer

    CategoryInput:
      type: object
      required: [slug, label, icon]
      properties:
        slug:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    CategoryUpdate:
      type: object
      properties:
        slug:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer

    CategoryStat:
      type: object
      required: [category, label, icon, count]
      properties:
        category:
          type: string
        label:
          type: string
        icon:
          type: string
        color:
          type: ["string", "null"]
        count:
          type: integer

    StoreStats:
      type: object
      required: [totalProducts, discordProducts, ottProducts, gamingProducts, inStockProducts]
      properties:
        totalProducts:
          type: integer
        discordProducts:
          type: integer
        ottProducts:
          type: integer
        gamingProducts:
          type: integer
        inStockProducts:
          type: integer

    StoreSettings:
      type: object
      required: [storeName, tagline]
      properties:
        storeName:
          type: string
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        discordLink:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        announcementText:
          type: ["string", "null"]
        announcementActive:
          type: boolean
        aboutText:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        discordSupportServer:
          type: ["string", "null"]
        discordShopLink:
          type: ["string", "null"]
        telegramLink:
          type: ["string", "null"]
        instagramLink:
          type: ["string", "null"]
        youtubeLink:
          type: ["string", "null"]
        discordWebhookUrl:
          type: ["string", "null"]

    StoreSettingsUpdate:
      type: object
      properties:
        storeName:
          type: string
          minLength: 1
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        discordLink:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        announcementText:
          type: ["string", "null"]
        announcementActive:
          type: boolean
        aboutText:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        discordSupportServer:
          type: ["string", "null"]
        discordShopLink:
          type: ["string", "null"]
        telegramLink:
          type: ["string", "null"]
        instagramLink:
          type: ["string", "null"]
        youtubeLink:
          type: ["string", "null"]
        discordWebhookUrl:
          type: ["string", "null"]

    AdminCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean