# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * GoluStore API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  label: string;
  icon: string;
  color?: string;
  description?: string;
  isActive?: boolean;
  sortOrder?: number;
}